% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_methods.R
\name{param_table}
\alias{param_table}
\title{Parameter table}
\usage{
param_table(x)
}
\arguments{
\item{x}{Object of type `hmm`.}
}
\description{
Helper function for other methods that prints parameters of `hmm` class in a human-readable format.
}
\keyword{internal}
