% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.hmm}
\alias{plot.hmm}
\title{Plotting Hidden Markov Models.}
\usage{
\method{plot}{hmm}(
  model,
  legend_position = "topright",
  xlab = "Time",
  ylab = "Observed emissions",
  cols = 1:model$m,
  ...
)
}
\arguments{
\item{model}{Class `hmm` model.}

\item{legend_position}{Position of legend in plots.}

\item{xlab}{Text for x-axis. Default 'Time'.}

\item{ylab}{Text for y-axis. Default 'Observed emissions'.}

\item{...}{}
}
\value{
Yields four plots stored in a list:

|              |                                                                   |
|--------------|-------------------------------------------------------------------|
| `p1`         | Plot of observed emissions by most likely hidden state (Global)   |
| `p2`         | Plot of observed emissions by most likely hidden state (Local)    |
| `p3`         | Plot of probs. used in Viterbi algorithm (Global)                 |
| `p4`         | Plot of state probabilities (Local)                               |
}
\description{
Plotting Hidden Markov Models.
}
\examples{
## Continuation of Earthquake example
\dontshow{example(hmm)}
plot(hmm.EQ)

}
